Fix occasional false failures in test-auto-upgrade pipeline job
* The `test-auto-upgrade` pipeline job failed once in the master branch with the following error.

  ```sh
  ERROR : [diff autoupgrade.TQG02-0_nocomment-082.sql.ref autoupgrade.TQG02-0_nocomment-082.sql.log] returned non-zero diff. See autoupgrade.TQG02-0_nocomment-082.sql.diff for details
  ERROR :   --> It is likely that bumping up FMT_PLAN_DEFINITION would fix such failures
  ```

  Below was the actual diff corresponding to the above failure.

  ```diff
  $ cat autoupgrade.TQG02-0_nocomment-082.sql.diff
  7,8d6
  < Burn|
  < Burn|Cavatelli
  29a28
  > Burn|Lasagna
  36a36
  > Burn|Pappardelle
  ```

* Below is the corresponding query file which has all comments removed (hence the `nocomment` in the sql file name).

  ```sql
  $ cat TQG02-0_nocomment-082.sql
  SELECT ALL (SELECT ALL alias2.lastName FROM names4 alias2 ORDER BY alias2.lastName LIMIT 1) AS alias2, alias1.favoritePasta FROM names4  INNER JOIN names4 AS alias1 ON (((names4.favoritePasta > alias1.firstName)) OR NOT (names4.lastName > 'Cereal')) WHERE (('Lord' < names4.firstName) OR (NOT (names4.lastName > names4.lastName) AND (2 < names4.id))) LIMIT 58;
  ```

* And below is the corresponding original query which has the `-- rowcount-only-check` string at the end.

  ```sql
  $ cat TQG02-0_comment-082.sql
  SELECT ALL (SELECT ALL alias2.lastName FROM names4 alias2 ORDER BY alias2.lastName LIMIT 1) AS alias2, alias1.favoritePasta FROM names4  INNER JOIN names4 AS alias1 ON (((names4.favoritePasta > alias1.firstName)) OR NOT (names4.lastName > 'Cereal')) WHERE (('Lord' < names4.firstName) OR (NOT (names4.lastName > names4.lastName) AND (2 < names4.id))) LIMIT 58; -- rowcount-only-check
  ```

* In this case, because of the `-- rowcount-only-check`, when the `TQG02` subtest ran with the older build
  of Octo, it only did a check of the number of lines of output comparison. It did not compare the actual
  output. And the `test-auto-upgrade` job should also have done only a rowcount check and not a diff.

* Below are all files related to query number `*082*`.

  ```sh
  $ ls -1 *082*
  TQG02-0_nocomment-082.sql
  TQG02-0_comment-082.sql
  TQG02-0_nocomment-082.sql.psql.out
  TQG02-0_nocomment-082.sql.unsorted.ref3
  TQG02-0_nocomment-082.sql.unsorted.log5
  TQG02-0_nocomment-082.sql.ref3
  TQG02-0_nocomment-082.sql.ref2
  TQG02-0_nocomment-082.sql.ref1
  TQG02-0_nocomment-082.sql.ref
  TQG02-0_nocomment-082.sql.octo.out
  TQG02-0_nocomment-082.sql.log5
  TQG02-0_nocomment-082.sql.log4
  TQG02-0_nocomment-082.sql.log3
  TQG02-0_nocomment-082.sql.log2
  TQG02-0_nocomment-082.sql.log1
  TQG02-0_nocomment-082.sql.log
  _ydboctoPCviGxL4q9XDSJCE8dL0l82.m
  _ydboctoPCviGxL4q9XDSJCE8dL0l82.o
  autoupgrade.TQG02-0_comment-082.sql.out
  autoupgrade.TQG02-0_nocomment-082.sql.ref.unsorted
  autoupgrade.TQG02-0_nocomment-082.sql.ref
  autoupgrade.TQG02-0_nocomment-082.sql.out
  autoupgrade.TQG02-0_nocomment-082.sql.log
  autoupgrade.TQG02-0_nocomment-082.sql.diff
  ```

* As can be seen, there is no file named `TQG02-0_nocomment-082.sql.diff`. This implies that the particular
  query had a `rowcount-only-check` specified. This is because `$fname.diff` gets created only in the
  `if` code path below (line 803) and not in the `else` code block.

  ```sh
   test_helpers.bash.in
   --------------------
    800         rowcountonlycheck=$(grep -c "rowcount-only-check" $query_file) || true
    801         if [[ $rowcountonlycheck -eq 0 ]]; then
    802                 # diff result is captured to stop_rocto before issuing a test fail
    803                 diff $fname.ref $fname.log >& $fname.diff || true
    804                 if [[ -s $fname.diff ]]; then
    805                         run_query_stop_rocto
    806                         exit 1
    807                 fi
    808         else
    809                 # .sql file requests only rowcount check. No exact content check.
    810                 # Check only # of lines in output match. Not contents of lines.
    811                 psqllines=$(wc -l $fname.ref | awk '{print $1}')
    812                 octolines=$(wc -l $fname.log | awk '{print $1}')
    813                 if [[ $psqllines -ne $octolines ]]; then
    814                         echo "Expected $psqllines lines but found $octolines lines in $fname.log"
    815                         run_query_stop_rocto
    816                         exit 1
    817                 fi
    818         fi
  ```

* Therefore, the `test-auto-upgrade` pipeline job script `tools/ci/build.sh` is also fixed to check for
  the existence of the `*.diff` file and if not do only a rowcount check (i.e. skip the diff).

+++ tools/ci/build.sh
@@ -630,13 +630,29 @@ FILE
 					else
 						cp $outfile $logfile
 					fi
-					difffile="autoupgrade.$sqlfile.diff"
-					diff $reffile $logfile > $difffile || true
-					if [[ -s $difffile ]]; then
-						echo "ERROR : [diff $reffile $logfile] returned non-zero diff. See $difffile for details" | tee -a ../errors.log
-						echo "ERROR :   --> It is likely that bumping up FMT_PLAN_DEFINITION would fix such failures" | tee -a ../errors.log
-						exit_status=1
-						errors_found=1
+					# Check if the output required only a rowcount check ("-- rowcount-only-check" in query file)
+					# This is deduced from the presence of *.diff files. If it is present, then a diff was done.
+					# If it is not, a rowcount check was done when the original test ran. Do the same thing below
+					# with the newer build of Octo.
+					if [[ -e $sqlfile.diff ]]; then
+						# .diff file exists. This means an actual diff was done. Do a diff in the new Octo build too.
+						difffile="autoupgrade.$sqlfile.diff"
+						diff $reffile $logfile > $difffile || true
+						if [[ -s $difffile ]]; then
+							echo "ERROR : [diff $reffile $logfile] returned non-zero diff. See $difffile for details" | tee -a ../errors.log
+							echo "ERROR :   --> It is likely that bumping up FMT_PLAN_DEFINITION would fix such failures" | tee -a ../errors.log
+							exit_status=1
+							errors_found=1
+						fi
+					else
+						# .diff file does not exist. This means only a rowcount check was done. Do the same check with the newer Octo build output.
+						oldoctolines=$(wc -l $reffile | awk '{print $1}')
+						newoctolines=$(wc -l $logfile | awk '{print $1}')
+						if [[ $oldoctolines -ne $newoctolines ]]; then
+							echo "ERROR : [$reffile has $oldoctolines lines but $logfile contains $newoctolines lines]" | tee -a ../errors.log
+							exit_status=1
+							errors_found=1
+						fi
 					fi
 				fi
 			done
